requirements.txt
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
groq>=0.11.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
httpx
pypdf
python-docx

